Move and mark internal classes
