+++ ironmq/src/main/scala/akka/stream/alpakka/ironmq/impl/IronMqPullStage.scala
@@ -2,23 +2,28 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.ironmq
+package akka.stream.alpakka.ironmq.impl
 
 import akka.Done
-import akka.stream.stage._
+import akka.annotation.InternalApi
 import akka.stream._
+import akka.stream.alpakka.ironmq.IronMqSettings.ConsumerSettings
+import akka.stream.alpakka.ironmq._
+import akka.stream.alpakka.ironmq.scaladsl.CommittableMessage
+import akka.stream.stage._
 
 import scala.concurrent.ExecutionContextExecutor
 import scala.util.{Failure, Success}
-import IronMqSettings.ConsumerSettings
-import akka.stream.alpakka.ironmq.scaladsl.CommittableMessage
 
-object IronMqPullStage {
+@InternalApi
+private[ironmq] object IronMqPullStage {
 
   private val FetchMessagesTimerKey = "fetch-messages"
 }
 
 /**
+ * Internal API.
+ *
  * This stage will fetch messages from IronMq and buffer them internally.
  *
  * It is implemented as a timed loop, each invocation will fetch new messages from IronMq if the amount of buffered
@@ -29,7 +34,9 @@ object IronMqPullStage {
  *
  * Keep in mind that the IronMq time unit is the second, so any value below the second is considered 0.
  */
-class IronMqPullStage(queue: Queue.Name, settings: IronMqSettings) extends GraphStage[SourceShape[CommittableMessage]] {
+@InternalApi
+private[ironmq] final class IronMqPullStage(queue: Queue.Name, settings: IronMqSettings)
+    extends GraphStage[SourceShape[CommittableMessage]] {
 
   import IronMqPullStage._
